commit 2.0 Changes in TaskProgramResponse

+++ src/main/java/org/example/kihelp/task/mapper/impl/TaskToTaskResponseMapperImpl.java
@@ -18,7 +18,7 @@ public class TaskToTaskResponseMapperImpl implements TaskToTaskResponseMapper {
                 subTitle,
                 task.getInfo(),
                 task.getTitle(),
-                (double)task.getPrice(),
+                task.getPrice(),
                 task.getDiscount(),
                 task.getType(),
                 task.getVisible(),
